+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Render item lists faster ([#1036](../../pull/1036))
 - Reduce bioformats source memory usage ([#1038](../../pull/1038))
 - Better pick the largest image from bioformats ([#1039](../../pull/1039), [#1040](../../pull/1040))
+- Cache histogram thresholds ([#1042](../../pull/1042))
 
 ## 1.19.3